+++ apx_sdk/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support EdDSA HTTP signatures.
+- Make `CoreType` and `get_core_type` public.
 
 ### Changed